Added container, game board, and boxes to css
The container made in html I styled with centering the game and gameboard. 
Then in the gameboard I made the boxes into columns & rows to lineup 5x5.
I ended the styling with giving my boxes borders and a more rectangular shape. After that I centered the text but I'm still having difficulty getting it to move from the top of the box to right in the middle